+++ src/kotae/KotaeAction.ts
@@ -8,8 +8,8 @@ import {
 } from '@slack/bolt'
 import { Logger, WebClient } from '@slack/web-api'
 import emojiRegex from 'emoji-regex'
-import { postMessage, postEphemeral, postInternalErrorMessage } from '../message/postMessage'
 import { OdaiUseCase } from '../odai/OdaiUseCase'
+import { postMessage, postEphemeral, postInternalErrorMessage } from '../slack/postMessage'
 import { getSlackUserList } from '../util/getSlackUserList'
 import { medalEmoji } from '../vote/util'
 import { KotaeUseCase } from './KotaeUseCase'
