+++ src/odai/OdaiAction.ts
@@ -2,7 +2,7 @@ import { App } from '@slack/bolt'
 import { KotaeUseCase } from '../kotae/KotaeUseCase'
 import { makePointRanking } from '../kotae/rank/makePointRanking'
 import { makeVotedCountRanking } from '../kotae/rank/makeVotedCountRanking'
-import { postMessage, postEphemeral, postInternalErrorMessage } from '../message/postMessage'
+import { postMessage, postEphemeral, postInternalErrorMessage } from '../slack/postMessage'
 import { getSlackUserList } from '../util/getSlackUserList'
 import { isImageUrl } from '../util/isImageUrl'
 import { VoteUseCase } from '../vote/VoteUseCase'
